Storage/Parallels.cpp: Fix harmless parfait integer overflow warning which is impossible, bugref:3409
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107424 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Storage/Parallels.cpp
@@ -695,7 +695,7 @@ static DECLCALLBACK(int) parallelsWrite(void *pBackendData, uint64_t uOffset, si
             Assert(uSector == 0);
             AssertMsg(pImage->cbFileCurrent % 512 == 0, ("File size is not a multiple of 512\n"));
             pImage->pAllocationBitmap[iIndexInAllocationTable] = (uint32_t)(pImage->cbFileCurrent / 512);
-            pImage->cbFileCurrent += pImage->PCHSGeometry.cSectors * 512;
+            pImage->cbFileCurrent += (uint64_t)pImage->PCHSGeometry.cSectors * 512;
             pImage->fAllocationBitmapChanged = true;
             uOffsetInFile = (uint64_t)pImage->pAllocationBitmap[iIndexInAllocationTable] * 512;
